Added commands for autosquash commenting (--fixup and --squash).
Added command to fixup a previous commit with informational message.

Added alias for fixup with comment to basically be thought of as a simple squash component, since it will ultimately not have a message after interactive rebase.

Added aliases similar to cis but to automatically commit all tracked changes.

Added cism/ciasm command to commit using --squash with message.